crud group member destination vote

+++ src/WebApi/Controllers/Admin/AdminGroupMemberController.cs
@@ -2,7 +2,10 @@ using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using WebApi.DTOs.Responses;
 using WebApi.DTOs.Responses.Admin.GroupMember;
+using WebApi.DTOs.Responses.Admin.GroupMemberDestinationVote;
+using WebApi.DTOs.Responses.GroupMemberDestinationVote;
 using WebApi.Persistence.Interfaces;
+using WebApi.Repositories.Implementations;
 using WebApi.Repositories.Interfaces;
 using WebApi.Security;
 
@@ -14,6 +17,7 @@ namespace WebApi.Controllers.Admin;
 public class AdminGroupMemberController(
     IUnitOfWork unitOfWork,
     IGroupRepository groupRepository,
+    IGroupMemberDestinationVoteRepository groupMemberDestinationVoteRepository,
     IGroupMemberRepository groupMemberRepository) : ControllerBase
 {
     [HttpGet]
@@ -41,8 +45,9 @@ public class AdminGroupMemberController(
         });
     }
 
-    [HttpGet("{id:guid}")]
-    public async Task<IActionResult> AdminGetGroupMemberById([FromRoute] Guid groupId, [FromRoute] Guid id)
+    [HttpGet("{memberId:guid}")]
+    public async Task<IActionResult> AdminGetGroupMemberById([FromRoute] Guid groupId, 
+        [FromRoute] Guid memberId)
     {
         var group = await groupRepository.GetGroupWithMembersAsync(groupId);
 
@@ -51,7 +56,7 @@ public class AdminGroupMemberController(
             return NotFound(new ErrorResponse("Group not found."));
         }
 
-        var groupMember = group.Members.SingleOrDefault(x => x.Id == id);
+        var groupMember = group.Members.SingleOrDefault(x => x.Id == memberId);
 
         if (groupMember is null)
         {
@@ -67,9 +72,9 @@ public class AdminGroupMemberController(
         });
     }
 
-    [HttpDelete("{id:guid}")]
+    [HttpDelete("{memberId:guid}")]
     public async Task<IActionResult> AdminRemoveGroupMemberById([FromRoute] Guid groupId,
-        [FromRoute] Guid id)
+        [FromRoute] Guid memberId)
     {
         var group = await groupRepository.GetGroupWithMembersAsync(groupId);
 
@@ -78,16 +83,53 @@ public class AdminGroupMemberController(
             return NotFound(new ErrorResponse("Group not found."));
         }
 
-        var userToRemove = group.Members.SingleOrDefault(m => m.Id == id);
+        var userToRemove = group.Members.SingleOrDefault(m => m.Id == memberId);
 
         if (userToRemove is null)
         {
             return NotFound(new ErrorResponse("The user is not a member of this group."));
         }
 
+        if (userToRemove.IsOwner)
+        {
+            return BadRequest(new ErrorResponse("It is not possible to remove the owner of group."));
+        }
+
         groupMemberRepository.Remove(userToRemove);
         await unitOfWork.SaveAsync();
 
         return NoContent();
+    }
+
+    [HttpGet("{memberId:guid}/destination-votes")]
+    public async Task<IActionResult> AdminGetGroupMemberAllDestinationVotesById([FromRoute] Guid groupId,
+        [FromRoute] Guid memberId, [FromQuery] int pageNumber = 1, [FromQuery] int pageSize = 10)
+    {
+        var group = await groupRepository.GetGroupWithMembersAsync(groupId);
+
+        if (group is null)
+        {
+            return NotFound(new ErrorResponse("Group not found."));
+        }
+
+        var isGroupMember = group.Members.Any(m => m.Id == memberId);
+
+        if (!isGroupMember)
+        {
+            return NotFound(new ErrorResponse("The user is not a member of this group."));
+        }
+
+        var (votes, hits) = await groupMemberDestinationVoteRepository.GetByMemberIdAsync(memberId,
+            pageNumber, pageSize);
+
+        return Ok(new AdminGetGroupMemberAllDestinationVotesByIdResponse
+        {
+            Data = votes.Select(x => new AdminGetGroupMemberAllDestinationVotesByIdResponseData
+            {
+                Id = x.Id,
+                CreatedAt = x.CreatedAt
+            }),
+            Hits = hits
+        });
     }
 }
